daemon/pyproject.toml
[project]
name = "daemon"
version = "0.1.0"
description = ""
authors = [
    {name = "Simon ",email = "175155691+Simon-Weij@users.noreply.github.com"}
]
readme = "../README.md"
requires-python = ">=3.13"
dependencies = [
    "evdev-binary (>=1.9.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "daemon", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "daemon.app:main"
